Added Desk! And handles, yippee!
Desk currently has a few disconnected surfaces on the top and in the seating spot towards the center. Unsure how to merge verticies. I've a desire to round the edges of the table, and will take that up shortly. Had a good time making the handles and figuring out the different rotation methods. Insets and ornamentation past the current inset incoming